Add test to for file export
